+++ inst/doc/rankinma_vignette.R
@@ -1,4 +1,4 @@
-## ---- include = FALSE---------------------------------------------------------
+## ----include = FALSE----------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
@@ -11,34 +11,7 @@ knitr::include_graphics("rankinma_logo.png")
 ## ----setup--------------------------------------------------------------------
 library(rankinma)
 
-## ---- eval = FALSE------------------------------------------------------------
-#  data <- data.frame(tx = c("A", "B", "C", "A", "B", "C"),
-#                     outcome = c("mortality", "mortality", "mortality", "recurrent", "recurrent", "recurrent"),
-#                     SUCRA = c(0.8, 0.7, 0.5, 0.9, 0.5, 0.8))
-
-## ---- eval = FALSE------------------------------------------------------------
-#  dataRankinma <- SetMetrics(data,
-#                             tx = tx,
-#                             outcome = outcome,
-#                             metrics = SUCRA,
-#                             metrics.name = "SUCRA")
-
-## ---- eval = FALSE------------------------------------------------------------
-#  PlotBar(data = dataRankinma)
-
-## ---- echo = FALSE, warning = FALSE, results = FALSE, message = FALSE, error = FALSE, fig.keep = "all", fig.cap = c("", "Figure 1. examples of bar chart for SUCRA on two outcomes."), fig.height = 5, fig.width = 5, fig.align = "center", out.width = "50%"----
-library("rankinma")
-data <- data.frame(tx = c("A", "B", "C", "A", "B", "C"),
-                   outcome = c("mortality", "mortality", "mortality", "recurrent", "recurrent", "recurrent"), 
-                   SUCRA = c(0.8, 0.7, 0.5, 0.9, 0.5, 0.8))
-dataRankinma <- SetMetrics(data, 
-                           tx = tx, 
-                           outcome = outcome, 
-                           metrics = SUCRA, 
-                           metrics.name = "SUCRA")
-PlotBar(data = dataRankinma)
-
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  library(netmeta)
 #  data(Senn2013)
 #  nmaOutput <- netmeta(TE,
@@ -49,7 +22,7 @@ PlotBar(data = dataRankinma)
 #                       data = Senn2013,
 #                       sm = "SMD")
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  dataMetrics <- GetMetrics(nmaOutput,
 #                            outcome = "HbA1c.random",
 #                            prefer = "small",
@@ -57,17 +30,17 @@ PlotBar(data = dataRankinma)
 #                            model = "random",
 #                            simt = 1000)
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  dataRankinma <- SetMetrics(dataMetrics,
 #                             tx = tx,
 #                             outcome = outcome,
 #                             metrics.name = "Probabilities")
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  PlotLine(data = dataRankinma,
 #           compo = TRUE)
 
-## ---- eval = TRUE, echo = FALSE, warning = FALSE, results = FALSE, fig.cap = "Figure 2A. an exmaple of composite line chart for probabilities of treatments on each rank.", fig.height = 5, fig.width = 7, fig.align = "center", out.width = "90%"----
+## ----eval = TRUE, echo = FALSE, warning = FALSE, results = FALSE, fig.cap = "Figure 2A. an exmaple of composite line chart for probabilities of treatments on each rank.", fig.height = 5, fig.width = 7, fig.align = "center", out.width = "90%"----
 library(netmeta)
 data(Senn2013)
 nmaOutput <- netmeta(TE, 
@@ -90,7 +63,7 @@ dataRankinma <- SetMetrics(dataMetrics,
 PlotLine(data = dataRankinma, 
          compo = TRUE)
 
-## ---- eval = TRUE, echo = FALSE, warning = FALSE, results = FALSE, fig.cap = "**Figure 2B**. an example of accumulative bar chart for probabilities of treatments on each rank.", fig.height = 5, fig.width = 7, fig.align = "center", out.width = "90%"----
+## ----eval = TRUE, echo = FALSE, warning = FALSE, results = FALSE, fig.cap = "**Figure 2B**. an example of accumulative bar chart for probabilities of treatments on each rank.", fig.height = 5, fig.width = 7, fig.align = "center", out.width = "90%"----
 library(netmeta)
 data(Senn2013)
 nmaOutput <- netmeta(TE, 
@@ -113,7 +86,7 @@ dataRankinma <- SetMetrics(dataMetrics,
 PlotBar(data = dataRankinma, 
         accum = TRUE)
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  library(netmeta)
 #  data(Senn2013)
 #  nmaOutput <- netmeta(TE,
@@ -124,7 +97,7 @@ PlotBar(data = dataRankinma,
 #                       data = Senn2013,
 #                       sm = "SMD")
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  nmaRandom <- GetMetrics(nmaOutput,
 #                          outcome = "HbA1c.random",
 #                          prefer = "small",
@@ -138,20 +111,20 @@ PlotBar(data = dataRankinma,
 #                          model = "common",
 #                          simt = 1000)
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  dataMetrics <- rbind(nmaRandom, nmaCommon)
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  dataRankinma <- (dataMetrics,
 #                   tx = tx,
 #                   outcome = outcome,
 #                   metrics = P.score,
 #                   metrics.name = "P-score")
 
-## ---- eval = FALSE------------------------------------------------------------
+## ----eval = FALSE-------------------------------------------------------------
 #  PlotBeads(data = dataRankinma)
 
-## ---- eval = TRUE, echo = FALSE,  fig.cap = "Figure 3. an example of beading plot for P-score on two outcomes", fig.height = 4, fig.width = 8, fig.align = "center", out.width = "80%"----
+## ----eval = TRUE, echo = FALSE,  fig.cap = "Figure 3. an example of beading plot for P-score on two outcomes", fig.height = 4, fig.width = 8, fig.align = "center", out.width = "80%"----
 library(netmeta)
 data(Senn2013)
 nmaOutput <- netmeta(TE,